Finished migration from CosmosDB to MongoDB. - docker-compose.yml sets up a development environment

+++ scc2022-backend/src/scc/session/Session.java
@@ -1,7 +1,7 @@
 package scc.session;
 
 public class Session {
-    public static final int VALIDITY_SECONDS = 3600;
+    public static final long VALIDITY_SECONDS = 3600L;
     public static final String COOKIE_NAME = "scc:session";
     private String nickname;
     private String cookieId;
